確認メール送信処理を別ファイルに切り出し / confirmationトークンの発行
